Refactor enum
