+++ src/power_grid_model_io/enum.py
