+++ tests/unit/converters/test_pandapower_converter_input.py
@@ -23,12 +23,12 @@ from power_grid_model import (
     initialize_array,
 )
 
-from power_grid_model_io._enum import _PandapowerTable as PpTable
 from power_grid_model_io.converters.pandapower_converter import (
     PP_COMPATIBILITY_VERSION_3_2_0,
     PandaPowerConverter,
     get_loss_params_3ph,
 )
+from power_grid_model_io.enum import _PandapowerTable as PpTable
 from tests.utils import MockDf, MockFn, assert_struct_array_equal
 
 type PandaPowerNet = pp.pandapowerNet
